refactor Rest layer
